[finishes #163694406] rename and create new views and models files
